# Improve CSV performance