@kou 2018-03-18
csv's upstream is https://github.com/ruby/csv .
Please change csv in upstream instead of the Ruby's repository.